extract state to controller and clean up
