+++ lib/containers/github-tab-header-container.js
@@ -2,19 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {QueryRenderer, graphql} from 'react-relay';
 
-import {incrementCounter} from '../reporter-proxy';
-import {
-  RemotePropType, RemoteSetPropType, BranchSetPropType, OperationStateObserverPropType, EndpointPropType,
-} from '../prop-types';
+import {EndpointPropType,} from '../prop-types';
 import RelayNetworkLayerManager from '../relay-network-layer-manager';
 import {UNAUTHENTICATED, INSUFFICIENT} from '../shared/keytar-strategy';
-import RemoteController from '../controllers/remote-controller';
 import ObserveModel from '../views/observe-model';
-import LoadingView from '../views/loading-view';
-import QueryErrorView from '../views/query-error-view';
-import GithubLoginView from '../views/github-login-view';
 import Author, {nullAuthor} from '../models/author';
-import GithubTabHeaderView from '../views/github-tab-header-view';
+import GithubTabHeaderController from '../controllers/github-tab-header-controller';
 
 export default class GithubTabHeaderContainer extends React.Component {
   static propTypes = {
@@ -22,11 +15,12 @@ export default class GithubTabHeaderContainer extends React.Component {
     loginModel: PropTypes.object.isRequired,
     endpoint: EndpointPropType.isRequired,
 
+    // Workspace
     currentWorkDir: PropTypes.string,
 
+    // Event Handlers
     handleWorkDirSelect: PropTypes.func,
     onDidChangeWorkDirs: PropTypes.func,
-    onDidUpdateRepo: PropTypes.func,
   }
 
   render() {
@@ -77,11 +71,15 @@ export default class GithubTabHeaderContainer extends React.Component {
     const {email, name, avatarUrl, login} = props.viewer;
 
     return (
-      <GithubTabHeaderView
+      <GithubTabHeaderController
+        user={new Author(email, name, login, false, avatarUrl)}
+
+        // Workspace
         currentWorkDir={this.props.workingDirectory}
-        committer={new Author(email, name, login, false, avatarUrl)}
-        handleWorkDirSelect={this.props.handleWorkDirSelect}
         getCurrentWorkDirs={this.props.getCurrentWorkDirs}
+
+        // Event Handlers
+        handleWorkDirSelect={this.props.handleWorkDirSelect}
         onDidChangeWorkDirs={this.props.onDidChangeWorkDirs}
       />
     );
@@ -89,11 +87,15 @@ export default class GithubTabHeaderContainer extends React.Component {
 
   renderNoResult() {
     return (
-      <GithubTabHeaderView
+      <GithubTabHeaderController
+        user={nullAuthor}
+        
+        // Workspace
         currentWorkDir={this.props.workingDirectory}
-        committer={nullAuthor}
-        handleWorkDirSelect={this.props.handleWorkDirSelect}
         getCurrentWorkDirs={this.props.getCurrentWorkDirs}
+
+        // Event Handlers
+        handleWorkDirSelect={this.props.handleWorkDirSelect}
         onDidChangeWorkDirs={this.props.onDidChangeWorkDirs}
       />
     );
