+++ lib/views/github-tab-header-view.js
@@ -6,50 +6,22 @@ import {AuthorPropType} from '../prop-types';
 
 export default class GithubTabHeaderView extends React.Component {
   static propTypes = {
-    currentWorkDir: PropTypes.string,
-    committer: AuthorPropType.isRequired,
+    user: AuthorPropType.isRequired,
 
+    // Workspace
+    workdir: PropTypes.string,
+    workdirs: PropTypes.arrayOf(PropTypes.string).isRequired,
+
+    // Event Handlers
     handleWorkDirSelect: PropTypes.func,
-    onDidChangeWorkDirs: PropTypes.func,
-    getCurrentWorkDirs: PropTypes.func.isRequired,
-  }
-
-  constructor(props) {
-    super(props);
-    this.state = {currentWorkDirs: []};
-  }
-
-  static getDerivedStateFromProps(props, state) {
-    return {
-      currentWorkDirs: props.getCurrentWorkDirs(),
-    };
-  }
-
-  componentDidMount() {
-    if (this.props.onDidChangeWorkDirs) {
-      this.disposable = this.props.onDidChangeWorkDirs(this.updateWorkDirs);
-    }
-  }
-
-  componentDidUpdate(prevProps) {
-    if (
-      prevProps.onDidChangeWorkDirs !== this.props.onDidChangeWorkDirs
-    ) {
-      if (this.disposable) {
-        this.disposable.dispose();
-      }
-      if (this.props.onDidChangeWorkDirs) {
-        this.disposable = this.props.onDidChangeWorkDirs(this.updateWorkDirs);
-      }
-    }
   }
 
   render() {
     return (
       <header className="github-Project">
-        {this.renderCommitter()}
+        {this.renderUser()}
         <select className="github-Project-path input-select"
-          value={this.props.currentWorkDir ? path.normalize(this.props.currentWorkDir) : undefined}
+          value={this.props.workdir ? path.normalize(this.props.workdir) : undefined}
           onChange={this.props.handleWorkDirSelect ? this.props.handleWorkDirSelect : () => {}}>
           {this.renderWorkDirs()}
         </select>
@@ -59,21 +31,15 @@ export default class GithubTabHeaderView extends React.Component {
 
   renderWorkDirs() {
     const workdirs = [];
-    for (const workdir of this.state.currentWorkDirs) {
+    for (const workdir of this.props.workdirs) {
       workdirs.push(<option key={workdir} value={path.normalize(workdir)}>{path.basename(workdir)}</option>);
     }
     return workdirs;
   }
 
-  updateWorkDirs = () => {
-    this.setState((state, props) => ({
-      currentWorkDirs: props.getCurrentWorkDirs(),
-    }));
-  }
-
-  renderCommitter() {
-    const login = this.props.committer.getLogin();
-    const avatarUrl = this.props.committer.getAvatarUrl();
+  renderUser() {
+    const login = this.props.user.getLogin();
+    const avatarUrl = this.props.user.getAvatarUrl();
 
     return (
       <img className="github-Project-avatar"
@@ -82,11 +48,5 @@ export default class GithubTabHeaderView extends React.Component {
         alt={`@${login}'s avatar`}
       />
     );
-  }
-
-  componentWillUnmount() {
-    if (this.disposable) {
-      this.disposable.dispose();
-    }
   }
 }
