+++ lib/views/github-tab-view.js
@@ -8,35 +8,40 @@ import {
 import LoadingView from './loading-view';
 import RemoteSelectorView from './remote-selector-view';
 import GithubTabHeaderContainer from '../containers/github-tab-header-container';
-import GithubTabHeaderView from './github-tab-header-view';
+import GithubTabHeaderController from '../controllers/github-tab-header-controller';
 import RemoteContainer from '../containers/remote-container';
 import {nullAuthor} from '../models/author';
 
 export default class GitHubTabView extends React.Component {
   static propTypes = {
+    rootHolder: RefHolderPropType.isRequired,
+
     // Connection
     loginModel: PropTypes.object.isRequired,
 
-    repository: PropTypes.object.isRequired,
+    // Workspace
     workspace: PropTypes.object.isRequired,
-    remoteOperationObserver: OperationStateObserverPropType.isRequired,
-    rootHolder: RefHolderPropType.isRequired,
-
     workingDirectory: PropTypes.string,
-    branches: BranchSetPropType.isRequired,
-    currentBranch: BranchPropType.isRequired,
+    getCurrentWorkDirs: PropTypes.func.isRequired,
+    changeWorkingDirectory: PropTypes.func.isRequired,
+
+    // Observers
+    remoteOperationObserver: OperationStateObserverPropType.isRequired,
+
+    // Remotes
     remotes: RemoteSetPropType.isRequired,
     currentRemote: RemotePropType.isRequired,
     manyRemotesAvailable: PropTypes.bool.isRequired,
+    isLoading: PropTypes.bool.isRequired,
+    branches: BranchSetPropType.isRequired,
+    currentBranch: BranchPropType.isRequired,
     aheadCount: PropTypes.number,
     pushInProgress: PropTypes.bool.isRequired,
-    isLoading: PropTypes.bool.isRequired,
 
+    // Event Handelrs
     handlePushBranch: PropTypes.func.isRequired,
     handleRemoteSelect: PropTypes.func.isRequired,
-    changeWorkingDirectory: PropTypes.func.isRequired,
     onDidChangeWorkDirs: PropTypes.func.isRequired,
-    getCurrentWorkDirs: PropTypes.func.isRequired,
   }
 
   render() {
@@ -63,15 +68,21 @@ export default class GitHubTabView extends React.Component {
           loginModel={this.props.loginModel}
           endpoint={this.props.currentRemote.getEndpoint()}
 
-          remoteOperationObserver={this.props.remoteOperationObserver}
-          pushInProgress={this.props.pushInProgress}
-          workingDirectory={this.props.workingDirectory}
+          // Workspace
           workspace={this.props.workspace}
+          workingDirectory={this.props.workingDirectory}
+
+          // Remote
           remote={this.props.currentRemote}
           remotes={this.props.remotes}
           branches={this.props.branches}
           aheadCount={this.props.aheadCount}
+          pushInProgress={this.props.pushInProgress}
 
+          // Observers
+          remoteOperationObserver={this.props.remoteOperationObserver}
+
+          // Event Handlers
           onPushBranch={() => this.props.handlePushBranch(this.props.currentBranch, this.props.currentRemote)}
         />
       );
@@ -109,19 +120,26 @@ export default class GitHubTabView extends React.Component {
           loginModel={this.props.loginModel}
           endpoint={this.props.currentRemote.getEndpoint()}
 
-          handleWorkDirSelect={e => this.props.changeWorkingDirectory(e.target.value)}
+          // Workspace
           currentWorkDir={this.props.workingDirectory}
           getCurrentWorkDirs={this.props.getCurrentWorkDirs}
+
+          // Event Handlers
+          handleWorkDirSelect={e => this.props.changeWorkingDirectory(e.target.value)}
           onDidChangeWorkDirs={this.props.onDidChangeWorkDirs}
         />
       );
     }
     return (
-      <GithubTabHeaderView
+      <GithubTabHeaderController
+        user={nullAuthor}
+
+        // Workspace
         currentWorkDir={this.props.workingDirectory}
-        committer={nullAuthor}
-        handleWorkDirSelect={this.props.handleWorkDirSelect}
         getCurrentWorkDirs={this.props.getCurrentWorkDirs}
+
+        // Event Handlers
+        handleWorkDirSelect={this.props.handleWorkDirSelect}
         onDidChangeWorkDirs={this.props.onDidChangeWorkDirs}
       />
     );
